Restore edit set functionality
