+++ fittrak-client/src/components/sets/AddSet.vue
@@ -79,6 +79,3 @@ export default {
   }
 };
 </script>
-
-<style scoped>
-</style>
